backend/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
alembic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
aiogram
httpx
bcrypt==4.0.1